### 2023-01-06 #5051 @myoelmy
Hi @kyleniemeyer, thanks for the update! Regarding the DOI issues I can confirm that all four of the found DOIs are correctly assigned and that the invalidity solely stems from the prefixes. Should I change the DOIs accordingly?
