### 2023-01-06 #5051 @jgostick
Hi @kyleniemeyer I know I am not editing anything at the moment, so am dead weight, but I really don't think I'll be be able to edit this one effectively. I don't even know what quantum friction is, and I don't C++. I am saying no to ones that are way outside my expertise since I had a review go sideways last summer that I was unable to deal with.
