### 2023-02-02 #5051 @lucydot
@kyleniemeyer sorry for my late reply - I completely missed this invite.

It looks very interesting! I'm not familiar with this research topic (or C++), but have a solid state theory background so not a million miles away. Happy to edit if someone better suited not on the team!
