### 2023-02-02 #5051 @kyleniemeyer
> Hi @kyleniemeyer, thanks for the update! Regarding the DOI issues I can confirm that all four of the found DOIs are correctly assigned and that the invalidity solely stems from the prefixes. Should I change the DOIs accordingly?

@myoelmy sorry for not responding to this before. Yes, please remove the DOI prefixes in the `.bib` file, the `doi` field should just have the actual DOI, and also add any missing DOIs. Thanks!
