### 2023-02-16 #5160 @editorialbot
[REVIEW]: QuaCa: an open-source library for fast calculations of steady-state quantum friction
**Submitting author:** @myoelmy (<a href="http://orcid.org/0000-0002-0120-9733">Marty Oelschläger</a>)
**Repository:** https://github.com/QuaCaTeam/quaca
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
