+++ paper/paper.bib
@@ -76,7 +76,7 @@
   volume  = {338},
   number  = {16},
   pages   = {1096-1104},
-  doi     = {https://doi.org/10.1002/andp.19103381603},
+  doi     = {10.1002/andp.19103381603},
   eprint  = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/andp.19103381603},
   file    = {einstein10.pdf:einstein10.pdf:PDF},
   url     = {https://onlinelibrary.wiley.com/doi/abs/10.1002/andp.19103381603},
@@ -320,17 +320,16 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
 @Article{oelschlaeger2021,
   author     = {Oelschläger, M. and Reiche, D. and Egerland, C. H. and Busch, K. and Intravaia, F.},
   journal    = {arXiv:2110.13635 [quant-ph]},
-  title      = {Electromagnetic {Viscosity} in {Complex} {Structured} {Environments}: {From} black-body to {Quantum} {Friction}},
-  year       = {2021},
-  month      = oct,
-  note       = {arXiv: 2110.13635},
+  title      = {Electromagnetic viscosity in complex structured environments: From blackbody to quantum friction
+},
+  year       = {20212},
+  month      = nov,
   abstract   = {We investigate the nonconservative open-system dynamics of an atom in a generic complex structured electromagnetic environment at temperature \$T\$. In such systems, when the atom moves along a translation-invariant axis of the environment, a frictional force acts on the particle. The effective viscosity due to friction results from the nonequilibrium interaction with the fluctuating (quantum) electromagnetic field, which effectively sets a privileged reference frame. We study the impact of both quantum or thermal fluctuations on the interaction and highlight how they induce qualitatively different types of viscosity, i.e. quantum and black-body friction. To this end, we develop a self-consistent non-Markovian description that contains the latter as special cases. In particular, we show how the interplay between the nonequilibrium dynamics, the quantum and the thermal properties of the radiation, as well as the confinement of light at the vacuum-material interface is responsible for several interesting and intriguing features. Our analyses is relevant for a future experimental test of noncontact friction and the resulting electromagnetic viscosity.},
-  annote     = {Comment: 16 pages, 8 figures},
-  file       = {:oelschlaeger22.pdf:PDF},
   keywords   = {Quantum Physics},
+  doi        = {10.1103/physreva.106.052205},
   shorttitle = {Electromagnetic {Viscosity} in {Complex} {Structured} {Environments}},
-  url        = {http://arxiv.org/abs/2110.13635},
-  urldate    = {2022-05-02},
+  url        = {https://journals.aps.org/pra/abstract/10.1103/PhysRevA.106.052205},
+  urldate    = {2023-02-17},
 }
 
 @Article{reiche2019,
@@ -373,7 +372,7 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
   title  = {Long-time Correlations in Nonequilibrium Dispersion Forces},
   school = {Humboldt-Universität zu Berlin, Mathematisch-Naturwissenschaftliche Fakultät},
   year   = {2021},
-  doi    = {http://dx.doi.org/10.18452/22412},
+  doi    = {10.18452/22412},
   file   = {reiche21.pdf:Volumes/reiche21.pdf:PDF;reiche21.pdf:reiche21.pdf:PDF},
 }
 
@@ -399,6 +398,7 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
   Journal                  = {Physics-Uspekhi},
[…]
