### 2023-02-20 #5051 @myoelmy
> Yes, I am available to review this submission.
>
> Also, I was just looking into the documentation and for some reason the katex script doesn't seem to render the equations. @myoelmy is there any way to fix this? I am guessing using kramdown or mathjax would help.

Should be fixed now.
