### 2023-02-24 QuaCaTeam/quaca#78 @arkajitmandal
[JOSS Review] Unit tests are failing
Here is the output when running "test_quaca_unit"

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
test_quaca_unit is a Catch v2.13.9 host application.
Run with -? for options

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

The operations calculate_tensor and the integrand_2d_k coincide
/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:75
...............................................................................

/Users/livi/Arkajit/quaca/test/UnitTests/GreensTensor/test_GreensTensorPlate_unit.cpp:140: FAILED:
  REQUIRE( approx_equal(Green_fancy_I_ct, Green_fancy_I_ik2d, "reldiff", 10E-4) )
with expansion:
  false

===============================================================================
test cases:   44 |   43 passed |  1 failed
assertions: 2723 | 2697 passed | 26 failed
