### 2023-02-24 QuaCaTeam/quaca#79 @arkajitmandal
[JOSS Review] Minor Issues
Some minor issue the authors can fix:
* The following link in the readme is returning 404: https://github.com/QuaCaTeam/quaca/blob/master/dev/testing.md
