### 2023-02-27 QuaCaTeam/quaca#78 @chegerland
Thank you for bringing up this issue!
Could you tell us more about how you have built the tests and what the environment is where you ran them?
In the [pipelines](https://github.com/QuaCaTeam/quaca/actions/workflows/build_and_test.yml) and on my computer no tests fail...
