### 2023-02-28 QuaCaTeam/quaca#78 @arkajitmandal
Hi Sorry about the late reply, I am using a macbook (intel) here is the details:

  Model Name:	MacBook Pro
  Model Identifier:	MacBookPro15,2
  Processor Name:	Quad-Core Intel Core i5
  Processor Speed:	2.3 GHz
  Number of Processors:	1
  Total Number of Cores:	4
  L2 Cache (per Core):	256 KB
  L3 Cache:	6 MB
  Hyper-Threading Technology:	Enabled
  Memory:	8 GB

When I do "cmake .. -DCMAKE_PREFIX_PATH=$CONDA_PREFIX" it shows:

-- Looking for sgemm_
-- Looking for sgemm_ - found
-- Configuring done
CMake Warning at app/Friction/CMakeLists.txt:2 (add_executable):
  Cannot generate a safe runtime search path for target Friction because
  there is a cycle in the constraint graph:

    dir 0 is [/Users/livi/Arkajit/quaca/lib]
    dir 1 is [/opt/anaconda3/envs/quaca-env/lib]
      dir 2 must precede it due to runtime library [libiomp5.dylib]
    dir 2 is [/opt/anaconda3/lib]
      dir 1 must precede it due to runtime library [libomp.dylib]

  Some of these libraries may not be found correctly.

CMake Warning at app/Decay/CMakeLists.txt:2 (add_executable):
  Cannot generate a safe runtime search path for target Decay because there
  is a cycle in the constraint graph:

    dir 0 is [/Users/livi/Arkajit/quaca/lib]
    dir 1 is [/opt/anaconda3/envs/quaca-env/lib]
      dir 2 must precede it due to runtime library [libiomp5.dylib]
    dir 2 is [/opt/anaconda3/lib]
      dir 1 must precede it due to runtime library [libomp.dylib]

  Some of these libraries may not be found correctly.

-- Generating done
-- Build files have been written to: /Users/livi/Arkajit/quaca/build

Could this be the issue ?
