### 2023-03-03 #5160 @vijaymocherla
Hi @lucydot,

I'm mostly done with my review. I didn't have any problems installing and testing the package on Linux environments. I've tested it on a few configurations and most didn't have any major issues that I couldn't easily resolve. But, I suggest the authors include a section in the documentation with known installation issues (like the one pointed out by @arkajitmandal).

Apart from that, I just have a few comments and suggestions for the authors. Also, this is the first time I'm reviewing a software paper for JOSS, so please let me know if I'm missing on any counts.

## Comments
- Modular structure
    + Since most scientific software is modular, I would request the author to elaborate more on the modularity of QuaCa and discuss any intentional choices made from a software design perspective to make the package extensible. This was briefly mentioned in lines 67-73, but a more clear justification would be helpful.

- Optimized version $F_\text{fric}$
    + Though the authors have cited `MEEP`, I would like the authors to clearly describe the current state of other open-source packages and the exact gap that QuaCa is trying to address.
    + Specifically, they have mentioned that QuaCa uses a numerically optimized version of $F_\text{fric}$. I would request the authors briefly describe the approach that is unique to QuaCa when compared with other FDTD packages that offer subroutines to calculate the Casimir force.
    + Additionally, I would request the authors to comment on the applicability of the implemented optimizations and limitations of the same, if any.
- Performance
    + It is mentioned that QuaCa is meant for efficient computation of $F_\text{fric}$ on any regular personal computer, but providing a few performance benchmarks and minimum prescribed hardware configuration would be ideal.

## Suggestions
+ To make it more accessible, I would recommend the authors consider improvising the documentation to provide more detailed information on creating the input file and reduce the code snippets (C++ class structures) in their documentation. Additionally, they could consider adding more information about arguments of their class methods.
+ I would also suggest setting a default value for the no. of threads (`OMP_NUM_THREADS`) to `1` (when no environment variable is set) within the package to prevent high CPU usage or oversubscription of processors which could risk crashing the system.
