### 2023-03-04 QuaCaTeam/quaca#78 @chegerland
I have tried it myself on a mac and unfortunately I do not get the same warnings/errors as you.
From the messages I understand that there are multiple versions of the openmp library.

I am very honest: I also do not understand why, if the above is true, cmake allows to compile the project anyways and why that would subsequently leed to one (!) of the tests failing.

If possible could you try one more thing for us please: Do not use the conda environment, but rather install the needed dependencies directly (I used homebrew for this).
