### 2023-03-05 QuaCaTeam/quaca#78 @arkajitmandal
I have tried now in several macos's and tried hundreds of times to install this without any success. To me it seems that its a issue related to that "AppleClang on MacOSX does not work well with OpenMP" : https://github.com/lczech/gappa/issues/15#issuecomment-769370615

Now I am getting an error:
[ 36%] Building CXX object src/CMakeFiles/quaca.dir/Polarizability/Polarizability.cpp.o
In file included from /Users/livi/Arkajit/quaca/src/Polarizability/Polarizability.cpp:7:
In file included from /Users/livi/Arkajit/quaca/src/Polarizability/../GreensTensor/GreensTensorFactory.h:4:
In file included from /Users/livi/Arkajit/quaca/src/Polarizability/../GreensTensor/GreensTensor.h:4:
In file included from /Users/livi/Arkajit/quaca/src/Polarizability/../GreensTensor/../Calculations/Integrations.h:3:
/usr/local/include/armadillo:104:15: warning: WARNING: use of OpenMP disabled; omp.h header not found [-W#pragma-messages]
      #pragma message ("WARNING: use of OpenMP disabled; omp.h header not found")
              ^
1 warning generated.
[ 38%] Building CXX object src/CMakeFiles/quaca.dir/PowerSpectrum/PowerSpectrum.cpp.o
In file included from /Users/livi/Arkajit/quaca/src/PowerSpectrum/PowerSpectrum.cpp:1:
In file included from /Users/livi/Arkajit/quaca/src/PowerSpectrum/PowerSpectrum.h:4:
In file included from /Users/livi/Arkajit/quaca/src/PowerSpectrum/../GreensTensor/GreensTensor.h:4:
In file included from /Users/livi/Arkajit/quaca/src/PowerSpectrum/../GreensTensor/../Calculations/Integrations.h:3:
/usr/local/include/armadillo:104:15: warning: WARNING: use of OpenMP disabled; omp.h header not found [-W#pragma-messages]
      #pragma message ("WARNING: use of OpenMP disabled; omp.h header not found")
              ^

It would be great if you could share your steps for installing quaca on a mac (also please share if your machine is a M1/M2 or a intel). Did you add any specific flags for the cmake ?

I am also attaching a pdf that show the steps taken to install and to get to the error:  [quaca.pdf](https://github.com/QuaCaTeam/quaca/files/10892873/quaca.pdf)

I am trying very hard to get a linux machine -- so that I can finish the rest of the review.
