### 2023-03-10 #5160 @arkajitmandal
Hi @lucydot, I have checked the software and used it following the documentation provided by the authors.
The package is well written and it is undoubtedly an important contribution to the field. Presently I have the following comment that the authors should address (in addition to those mentioned by @vijaymocherla) before I can recommend this work for publication.

* **Installation**: Presently package is nearly very difficult to install on a mac (I have tried three macs one with M1, two with intel) in all of them I have faced issued in installing. However such an issue do not arise in a linux (I tried this in Ubuntu). I think the authors have two options: (1) modify cmake to make sure the package is installable in mac (or (2) To clearly state that they do not support mac in the readme. On mac due to the problem of installation (most likely due to some issue with openmp) sometimes when it is installed the many of the test fail. The tests pass on Ubuntu.
https://github.com/QuaCaTeam/quaca/issues/78#issue-1598013076

* **Readme**: For installing the package on linux, I see that the authors use the following line in workflows/build_and_test.yml
```
sudo apt-get update && sudo apt-get install libblas-dev liblapack-dev cmake make g++ libboost-filesystem-dev libboost-program-options-dev
```
which is very convenient. In the readme however they provide an alternate route with conda. I would suggest the author to also add this alternate installation approach in the readme.

* **Minor Issue** : Some links do not work:
https://github.com/QuaCaTeam/quaca/issues/79#issue-1598023878
