### 2023-03-13 #5160 @lucydot
Thank you very much @arkajitmandal and @vijaymocherla for your expertise, and timely reviews.

@myoelmy over to you! It looks like you have several points to act on, both from comment above ☝️ and earlier (https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277). I expect you will not be able to resolve everything in one sitting, but please do keep us updated with your work in this thread as you progress.
