### 2023-03-20 QuaCaTeam/quaca#80 @myoelmy
[JOSS Review] Comments on Modular structure and Optimized Version
https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277
