### 2023-03-20 QuaCaTeam/quaca#81 @myoelmy
[JOSS Review] Performance
This issue shall resolve the comments from https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277:

-Performance

It is mentioned that QuaCa is meant for efficient computation of
 on any regular personal computer, but providing a few performance benchmarks and minimum prescribed hardware configuration would be ideal.

[...]

I would also suggest setting a default value for the no. of threads (OMP_NUM_THREADS) to 1 (when no environment variable is set) within the package to prevent high CPU usage or oversubscription of processors which could risk crashing the system.
