### 2023-03-21 QuaCaTeam/quaca#80 @ddanzer
This issue shall resolve the following comments from https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277:

- Modular structure

- [ ]   Since most scientific software is modular, I would request the author to elaborate more on the modularity of QuaCa and discuss any intentional choices made from a software design perspective to make the package extensible. This was briefly mentioned in lines 67-73, but a more clear justification would be helpful.

- Optimized version

- [ ] Though the authors have cited MEEP, I would like the authors to clearly describe the current state of other open-source packages and the exact gap that QuaCa is trying to address.

- [ ] Specifically, they have mentioned that QuaCa uses a numerically optimized version of $F_{\mathrm fric}$. I would request the authors briefly describe the approach that is unique to QuaCa when compared with other FDTD packages that offer subroutines to calculate the Casimir force.

- [ ] Additionally, I would request the authors to comment on the applicability of the implemented optimizations and limitations of the same, if any.
