### 2023-03-21 25c6265
Update paper.md
Implemented replies to comments in issue https://github.com/QuaCaTeam/quaca/issues/80#
