### 2023-03-21 QuaCaTeam/quaca#80 @ddanzer
Concerning point 1.

We replaced the sentence in line 67

	"On the other hand (ii), we have chosen an implementation strategy where any of the physical quantities, such as the permittivity, the polarizability, the Green tensor or the power spectrum, are assigned to individual objects in the package. These can be tested, replaced, computed or used independently. By virtue of this modular structure, the QuaCa package can be easily adapted to different geometries and materials, parts of the code can be used as a library performing a subroutine in a larger project, or it can even be extended to compute other observables of fluctuation-induced light-matter interactions."

with

	"On the other hand (ii), we have chosen a modular implementation strategy guided by physical principles. We assigned each meaningful physical observable in the setup to individual objects in the package which can be tested, replaced, computed or used independently. Irrespective of the particular intricacies that come with the calculation of quantum friction (i), the code structure reflects the individual building blocks of the physical setup, i.e. the atom (polarizability), the electromagnetic vacuum field in the presence of the given geometry (Green tensor) as well as the respective material properties (permittivity). By virtue of this modular structure, the QuaCa package can be easily adapted to different geometries and materials in the calculation of quantum friction, (physically relevant) objects of the code can be used as a library performing a subroutine in a larger project, or they can be used to compute other observables of fluctuation-induced light-matter interactions."
