### 2023-03-21 QuaCaTeam/quaca#80 @ddanzer
Concerning point 2.

We are not aware of any open-source software that is designed to compute the steady-state atom-surface quantum friction, other than brute force implementations of the coupled integrals occurring in $F_{\mathrm{fric}}$. As explained in lines 60-66, previous code projects that focus on _other_ fluctuation-induced effects are, to the best of our knowledge, not directly applicable to the situation.
