### 2023-03-21 QuaCaTeam/quaca#80 @ddanzer
Concerning point 3.

As in point 2., other open-source software can not, we think, be directly applied to the situation of steady-stae atom-surface quantum friction. The term "optimized" in the draft refers to a particular analytical form of $F_{\mathrm{fric}}$, "where some nested integrals decouple" (line 64), which allows for an optimized performance with respect to a brute-force implementation. We acknowledge that this can lead to a misunderstanding. To make it more clear, we have modified the sentence in line 64 which now reads:

	"Here, QuaCa uses an analytically equivalent version of $F_{\mathrm{ fric}}$, which is particularly suited for numerical computations. We decouple nested integrals and take explicit care of the occurring poles. This allows for an efficient and reliable computation of the force. For details of the procedure, we refer to Ref. [@oelschlager2019]."
