### 2023-03-21 QuaCaTeam/quaca#80 @ddanzer
Concerning point 4.

The in point 3 (and the paper) described "optimizations" have made a numerical calculation of $F_{\mathrm{fric}}$, in the situation described in the paper, possible in the first place. By implication, the QuaCa package is somewhat limited to an efficient calculation of steady-state quantum friction. Due to the modular structure driven by physical arguments, we may regain some flexibility, since many other fluctuation-induced forces can be computed from the same building blocks. We try to demonstrate this flexibility by additionally providing an executable for computing atomic decays rates. However, we have no intention to be in competition with other open-source software here. The strength of the package is the computation of (nonequilibrium) steady-state atom-surface quantum friction.
