+++ docs/dev/testing.md
@@ -11,21 +11,11 @@ In order to ensure a solid quality of QuaCa we thoroughly implemented unit test 
    * For the [vacuum Green's tensor](api/greenstensor?id=greenstensorvacuum) in the limit of $v\ll c$ we find
     - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2}  \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\} \sim   \mathrm{diag}[1,1,1]\frac{\omega^3}{6\pi \epsilon_0 c^3}$
 
-    - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2}  \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\}k_x \sim \frac{v/c }{6\pi\epsilon_0} 
- \frac{ \omega^4 }{c^4}   
-\mathrm{diag}\left[
-    1,\,
-    2
-    ,\,
-     2\right]$
+    - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2} \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\} k_x \sim \frac{v/c}{6\pi\epsilon_0}  \frac{ \omega^4 }{c^4} \mathrm{diag}\left[ 1,\, 2 ,\, 2\right]$
 
-    - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2}  \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\}\left[1-\exp(\hbar\beta(\omega+\mathbf{k}^\intercal\mathbf{v}))\right]^{-1} \stackrel{\hbar\beta\omega\gg 1}\sim  -
-\frac{2\mathrm{diag}[1,2,2]}{15 \pi\epsilon_0 }\frac{v^2}{c^2}
-\frac{\omega^3}{c^3}\frac{1}{\hbar\beta\omega } $
+    - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2}  \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\}\left[1-\exp(\hbar\beta(\omega+\mathbf{k}^\intercal\mathbf{v}))\right]^{-1} \stackrel{\hbar\beta\omega\gg 1}\sim - \frac{2\mathrm{diag}[1,2,2]}{15 \pi\epsilon_0 }\frac{v^2}{c^2} \frac{\omega^3}{c^3}\frac{1}{\hbar\beta\omega } $
 
-    - $\int\frac{\mathrm{d}^2\mathbf{ k}}{(2\pi)^2}  \mathrm{Im}\left\{\underline{G}_0(\mathbf{ k}, z\to z',\omega+\mathbf{k}^\intercal\mathbf{v})\right\}\left[1-\exp(\hbar\beta(\omega+\mathbf{k}^\intercal\mathbf{v}))\right]^{-1}k_x \stackrel{\hbar\beta\omega\gg 1}\sim  -
- \frac{\mathrm{diag}\left[1,2,2\right]}{30\pi\epsilon_0}\frac{v}{c}\frac{\omega^4}{c^4}
- \frac{1}{\hbar\beta\omega} $
[…]
