### 2023-03-28 37d9ea1
Suggestion for a more compact and informative description of the interface of QuaCa
