+++ docs/api/greenstensor.md
@@ -2,49 +2,39 @@
 This is an abstract class that defines a Green's tensor.
 A specific kind of Green's tensor will be a child of this class.
 
-```cpp
-class GreensTensor {
-protected:
-  double v; //velocity of the particle 
-  double beta; // inverse temperature
+## Member functions
 
-public:
-  // constructor
-  GreensTensor(double v, double beta);
-  explicit GreensTensor(const std::string &input_file);
+### `GreensTensor(std::string input_file)`
+Description: Input file constructor of the class.
+* Input parameters: 
+    - `std::string input_file`: Name of the input file
+* Return value: `void`
 
-  // calculate the tensor in frequency and momentum space
-  virtual void calculate_tensor(double omega, vec::fixed<2> k,
-                                cx_mat::fixed<3, 3> &GT) const = 0;
-
-  // integrate over a two-dimensional k space
-  virtual void integrate_k(double omega, cx_mat::fixed<3, 3> &GT,
-                           Tensor_Options fancy_complex,
-                           Weight_Options weight_function) const = 0;
-
-  // calculates and returns a characteristic frequency
-  virtual double omega_ch() const = 0;
-
-  // getter functions
-  double get_v() const { return this->v; };
-  double get_beta() const { return this->beta; };
-
-  // setter function
-  virtual void set_v(double v_new) { this->v = v_new; };
-};
-```
-
-### `# GreensTensor(std::string input_file)`
-Input file constructor of the class.
-
-### `# GreensTensor(double v, double beta)`
+### `GreensTensor(double v, double beta)`
 Direct constructor of the class.
+* Input parameters:
+    - `double v`: velocity of the atom
+    - `double beta`: inverse temperature of the system
+* Return value: `void`:
 
-### `# virtual void calculate_tensor(double omega, vec::fixed<2> k,cx_mat::fixed<3, 3> &GT) const = 0;`
+### `virtual void calculate_tensor(double omega, vec::fixed<2> k,cx_mat::fixed<3, 3> &GT) const = 0;`
 Calculates the Green's tensor, i.e. $\underline{G}(\mathbf{k},z,z', \omega + k_x v)$, where $x$ is the axis along the motion with velocity $v$, and puts the result into the matrix `GT`. Here, the $x$ and $y$ coordinate are calculated in reciprocal space $(x-x',y-y') \to (k_x,k_y)$, while the $z$ or $z'$ component remain in the space domain. The $z$ components are defined in the respective child class. For more physical details, see [this publication](https://journals.aps.org/pra/abstract/10.1103/PhysRevA.94.042114). 
+* Input parameters:
+    - `double omega`: Frequency, at which the Green's tensor is evaluated.
+    - `vec:fixed<2> k`: momentum vector $\mathbf{k} = (k_x, k_y)$.
+    - `cx_mat:fixed<3, 3> &GT`: 3x3 matrix, where the resulting Green's tensor is stored.
+* Return value: `void`
 
-###  `# virtual void integrate_k(double omega, cx_mat::fixed<3, 3> &GT, Tensor_Options fancy_complex, Weight_Options weight_function) const = 0;`
+###  `virtual void integrate_k(double omega, cx_mat::fixed<3, 3> &GT, Tensor_Options fancy_complex, Weight_Options weight_function) const = 0;`
 Compute the integral over the momentum space of the Green's tensor i.e. $\int_{-\infty}^{\infty} \frac{\mathrm{d}^2 \mathbf{k}}{(2 \pi)^2} \; f(\mathbf{k},\omega) \underline{G}(\mathbf{k},z,z', \omega + k_x v)$, where $f(\mathbf{k},\omega)$ is some weight function.  The result is stored in `GT`.
+
+* Input parameters:
[…]
