### 2023-04-19 QuaCaTeam/quaca#78 @chegerland
We resolve this issue by stating that we do not support mac installations.
