### 2023-05-04 #5160 @ddanzer
Dear @lucydot @arkajitmandal @vijaymocherla ,

while the code and doc review is still in progress, you may want to have a first look on the changes we made in the manuscript (https://github.com/QuaCaTeam/quaca/tree/master/paper) concerning the points of https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277.
For additional review, you may use the open issue at https://github.com/QuaCaTeam/quaca/issues/80 which also contains a specific response to each of the raised points.

Thanks!
