### 2023-05-09 cf4a090
added small benchmark
