### 2023-05-09 79dcc06
added small benchmark
