### 2023-05-09 81caae2
Added the format for the documentation to the remaining files in docs/api
