### 2023-05-11 #5160 @ddanzer
Dear @lucydot @arkajitmandal @vijaymocherla,

we have added a benchmark and we have set the default value for the no. of threads (OMP_NUM_THREADS) to 1 in our most recent merge (https://github.com/QuaCaTeam/quaca/pull/83) replying to the correpsonding point in https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277.
For additional review, you may use the open issue at https://github.com/QuaCaTeam/quaca/issues/81.

Thanks!
