### 2023-05-15 #5160 @vijaymocherla
Dear @ddanzer,

Thanks for the update. I think this ([QuaCaTeam/quaca#80](https://github.com/QuaCaTeam/quaca/issues/80)) mostly answers some of the issues raised in my earlier review.

@lucydot I believe if the authors would like to emphasize the efficiency of their package and justify the publication of their software package, I would like them to provide some relevant performance benchmarks or any additional details substantiating the need for developing a stand-alone package like QuaCa.  Otherwise, I suggest the authors make a decision if they would like to present it as a software package that does atom-surface quantum-friction calculations. In that case, it would make QuaCa, a package that addresses a niche problem. This begs the question, what would justify the need for a standalone package for some niche calculations? Does JOSS have any editorial criteria in this regard? In that case, I would suggest the authors make necessary changes to their draft and QuaCa to comply with JOSS requirements.
