+++ paper/paper.md
@@ -25,13 +25,13 @@ affiliations:
    index: 2
  - name: Humboldt-Universität zu Berlin, Institut für Physik, 12489 Berlin, Germany
    index: 3
-date: 28 June 2022
+date: 25 May 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-QuaCa is an extensible library facilitating the fast computation of steady-state atom-surface quantum friction.
+QuaCa is an extensible library facilitating the computation of steady-state atom-surface quantum friction.
 Due to its modular domain-driven structure, QuaCa can be of further use to calculate relevant quantities that are often needed in the context of electromagnetic dispersion forces.
 Quantum (or Casimir) friction is a quantum-optical fluctuation-induced force that occurs in dynamical nonequilibrium, i.e. when a number of bodies are moving relatively to one another [@pendry1997;@scheel2009].
 The frictional interaction between the moving system of interest and its environment is mediated by the (material-modified) quantum vacuum and persists even at zero temperature.
@@ -40,15 +40,16 @@ For a comprehensive review of quantum friction in various contexts as well as re
 
 # Statement of need
 
-Due to the relative weakness of quantum friction with respect to comparable fluctuation-induced effects in equilibrium, such as the Casimir(-Polder) or the van der Waals effect [@dalvit2011], quantum friction has, to the best of our knowledge, yet evaded experimental confirmation.
+Due to the relative weakness of quantum friction with respect to related fluctuation-induced effects in equilibrium, such as the Casimir(-Polder) or the van der Waals effect [@dalvit2011], quantum friction has yet evaded experimental confirmation.
 This has instigated a surge of interest in exploring potentially useful designs [@volokitin2011;@milton2016;@farias2020;@lombardo2021] (see also [@reiche2022] and references therein) and a sophisticated (numerical) optimization of the experimental setup appears necessary in order to promote the effect to the measurable realm [@oelschlager2019;@reiche2021]. 
-That is where QuaCa can come into use.
+This gap can be closed using QuaCa. 
+QuaCa pioneers the numerical simulation of quantum friction and, to date, remains the only ready-to-use and openly available package for exploring experimental designs that can be utilized to measure quantum friction.
+
 
 # Physics of quantum friction
 
-QuaCa computes the quantum frictional force experienced by a microscopic particle (e.g. atom or nano-particle) moving along the invariant direction of an ensemble of macroscopic bodies.
-We focus on the late-time regime where all the transitional dynamics has settled and the particle can be assumed to be in a nonequilibrium steady-state moving with constant non-relativistic velocity $v$.
-In the regime of linear coupling between system and environment and in accordance with Refs. [@intravaia2014;@intravaia2016;@intravaia2016a;@intravaia2019], the frictional force connected to the electric interaction reads
+We focus on the nonequilibrium steady-state of a microscopic particle moving with non-relativistic velocity $v$.
+The frictional force connected to the electric interaction then reads [@intravaia2014;@intravaia2016;@intravaia2016a;@intravaia2019]
 $$
   F_\mathrm{fric} =
 -2
@@ -57,10 +58,10 @@ $$
 \underline{S}^\intercal(hv-\omega)\underline{G}_\Im(h, \mathbf{R}_a, \omega)
 \right],
 $$
[…]
