### 2023-06-05 #5160 @ddanzer
Dear @lucydot, dear @vijaymocherla, dear @arkajitmandal,

we have modified the documentation https://github.com/QuaCaTeam/quaca/pull/82 following the suggestions of https://github.com/openjournals/joss-reviews/issues/5160#issuecomment-1453048277 . We believe to have addressed all of the open points now and are looking forward to further suggestions and comments.

Thank you very much for the thorough and helpful review.
