### 2023-06-05 #5160 @vijaymocherla
Thanks for the update @ddanzer !

I believe that's all from my side, I do not have anymore queries or suggestions.

Over to you @lucydot .

PS:
A little explanation to clarify my earlier comment, I certainly do not think using the term "niche" undermines the scope of your package. On the other hand "pioneering" would leave QuaCa with much bigger shoes to fill, which could go both ways. But, this is a decision I leave it for you to decide.
