### 2023-06-13 5dbc271
Update paper.md
For JOSS review process; editorial checks
