### 2023-06-13 #5160 @lucydot
Hi @ddanzer I've made some suggested changes to the manuscript and issued a PR (https://github.com/QuaCaTeam/quaca/pull/85).

Please also check that the capitalisation for the reference list is consistent : Capitalize only the first word of an article title. Capitalize proper nouns, initials, and acronyms in a title. For example, I can see that "Power, Force and Torque" is capitalised. Also "Plane-Sphere Geometry". Please check the other references for consistency.
