### 2023-06-15 #5160 @chegerland
Hello @lucydot I have merged your suggested changes, thank you very much!

Is it ok if we do not have consistent capitalization? I would rather not change the original titles of the references, since that is how you can best find them.
