### 2023-06-21 #5160 @lucydot
> Is it ok if we do not have consistent capitalization? I would rather not change the original titles of the references, since that is how you can best find them.

I think it should be ok. I may have misinterpreted the reference to `capitalization` in the checklist: thinking on this probably refers to something like `r programming language` instead of `R programming language`. However I will check with the editorial team and come back on it.
