### 2023-06-28 #5160 @lucydot
@chegerland @ddanzer I've discussed this with the wider editorial team. We would prefer consistency, with all titles in sentence case. We don't see an issue with locating original references, as all are attached to a doi. In certain cases the capitalisation should be maintained (for proper nouns) using brackets as you have done e.g. {R}, {P}ython, {B}ayesian, and chemical symbols.
