### 2023-06-28 74d7ede
fix capitalization on german references
