+++ paper/paper.bib
@@ -70,7 +70,7 @@
 
 @Article{einstein1910,
   author  = {Einstein, A. and Hopf, L.},
-  title   = {Über einen Satz der Wahrscheinlichkeitsrechnung und seine Anwendung in der Strahlungstheorie},
+  title   = {Über einen {Satz} der {Wahrscheinlichkeitsrechnung} und seine {Anwendung} in der {Strahlungstheorie}},
   journal = {Annalen der Physik},
   year    = {1910},
   volume  = {338},
@@ -241,7 +241,7 @@
 }
 
 @article{hartmann2020,
-  title = {{{CaPS}}: {{casimir effect}} in the {{plane}}-{{sphere geometry}}},
+  title = {{{CaPS}}: {{Casimir effect}} in the {{plane}}-{{sphere geometry}}},
   shorttitle = {{{CaPS}}},
   author = {Hartmann, Michael and Ingold, Gert-Ludwig},
   year = {2020},
@@ -461,7 +461,7 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
 
 @Article{milton2016,
   author   = {Milton, Kimball A. and Høye, Johan S. and Brevik, Iver},
-  title    = {The reality of casimir friction},
+  title    = {The reality of {Casimir} friction},
   journal  = {Symmetry},
   year     = {2016},
   volume   = {8},
@@ -511,7 +511,7 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
 
 @Article{milonni1981,
   author  = {Milonni,P. W.},
-  title   = {Quantum mechanics of the Einstein–Hopf model},
+  title   = {Quantum mechanics of the {Einstein}–{Hopf} model},
   doi     = {10.1119/1.12552},
   eprint  = {https://doi.org/10.1119/1.12552},
   number  = {2},
