+++ paper/paper.bib
@@ -166,7 +166,7 @@
 
 
 @Article{busch2011,
-  Title                    = {Discontinuous Galerkin methods in nanophotonics},
+  Title                    = {Discontinuous {Galerkin} methods in nanophotonics},
   Author                   = {Busch, K. and K\"onig, M. and Niegemann, J.},
   Journal                  = {Laser \& Photonics Reviews},
   Year                     = {2011},
@@ -431,7 +431,7 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
 
 @InBook{johnson2011,
   pages     = {175--218},
-  title     = {Numerical methods for computing casimir interactions},
+  title     = {Numerical methods for computing {Casimir} interactions},
   publisher = {Springer Berlin Heidelberg},
   year      = {2011},
   author    = {Johnson, Steven G.},
