### 2023-06-30 #5160 @chegerland
Hello @lucydot,
I have incorporated all the suggested changes and made a release ([Version 1.0](https://github.com/QuaCaTeam/quaca/releases/tag/v1.0)).
Then I uploaded all to [Zenodo](https://zenodo.org/record/8101846), which resulted in the DOI [10.5281/zenodo.8101845](https://doi.org/10.5281/zenodo.8101845).

I hope that covers all on the authors checklist and we are ready to go! 🎉
