### 2023-07-02 #5160 @lucydot
I’m away at a conference next week, and plan to follow up on JOSS submissions w/c 10th July. Seeing as this is so close, I will try to check-in next week, but I may get carried away with conference activities. Just a heads-up so you understand if there is a delay.
