+++ paper/paper.bib
@@ -147,6 +147,7 @@
 @phdthesis{oelschlager2019,
   title = {Fluctuation-induced phenomena in nanophotonic systems},
   author = {Oelschl{\"a}ger, M.},
+  Doi = {10.18452/21484},
   year = {2019}
 }
 
