### 2023-07-03 #5160 @chegerland
> @chegerland there is a typo in the title of the Zenodo release, please could you fix?
done!

> @chegerland can I check there is no DOI available for the referenced thesis? I can’t find one online but was under impression that most educational institutions upload to a institution repository and assign a corresponding doi.
very well spotted. I have added the DOI to the reference in [this commit](https://github.com/QuaCaTeam/quaca/commit/86dd683187af1c64b3c4b2607dbecad4d43a7fd1).

Thank you for the heads-up! 😁
