+++ paper/paper.md
@@ -10,6 +10,7 @@ tags:
 authors:
   - name: Marty Oelschläger
     affiliation: "1,2,3"
+    orcid: 0000-0002-0120-9733
   - name: Simon Hermann
     affiliation: 3
   - name: Christoph H. Egerland
