### 2023-07-13 d6db766
Adjust journal names so not abbreviated
Inconsistent references, with mixture of abbreviated and non-abbreviated journal names. Fixed in this commit.
