+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{intravaia18,
   title = {Quantum rolling friction},
   author = {Intravaia, F. and Oelschl\"ager, M. and Reiche, D. and Dalvit, D. A. R. and Busch, K.},
-  journal = {Phys. Rev. Lett.},
+  journal = {Physical Review Letters},
   volume = {123},
   issue = {12},
   pages = {120401},
@@ -15,7 +15,7 @@
 
 @Article{reiche2020a,
   author    = {Reiche, D. and Intravaia, F. and Hsiang, J.-T. and Busch, K. and Hu, B.-L.},
-  journal   = {Phys. Rev. A},
+  journal   = {Physical Review A},
   title     = {Nonequilibrium thermodynamics of quantum friction},
   year      = {2020},
   month     = {Nov},
@@ -120,7 +120,7 @@
   pages = {042114},
   doi = {10.1103/PhysRevA.94.042114},
   file = {/home/marty/Zotero/storage/GF7QXTDE/Intravaia et al_2016_Non-Markovianity in atom-surface dispersion forces.pdf},
-  journal = {Phys. Rev. A},
+  journal = {Physical Review A},
   keywords = {_tablet},
   number = {4}
 }
@@ -128,7 +128,7 @@
 @Article{intravaia2016,
   Title                    = {Failure of local thermal equilibrium in quantum friction},
   Author                   = {Intravaia, F. and Behunin, R. O. and Henkel, C. and Busch, K. and Dalvit, D. A. R.},
-  Journal                  = {Phys. Rev. Lett.},
+  Journal                  = {Physical Review Letters},
   Year                     = {2016},
 
   Month                    = {Sep},
@@ -160,7 +160,7 @@
   pages = {050101},
   doi = {10.1103/PhysRevA.89.050101},
   file = {/home/marty/Zotero/storage/7B7VZH65/Intravaia et al_2014_Quantum friction and fluctuation theorems.pdf;/home/marty/Zotero/storage/UMI5ZGS7/Intravaia et al_2014_Quantum friction and fluctuation theorems.pdf},
-  journal = {Phys. Rev. A},
+  journal = {Physical Review A},
   keywords = {_tablet},
   number = {5}
 }
@@ -202,7 +202,7 @@
 @Article{mkrtchian2003,
   author    = {Mkrtchian, Vanik and Parsegian, V. Adrian and Podgornik, Rudi and Saslow, Wayne M.},
   title     = {Universal thermal radiation drag on neutral objects},
-  journal   = {Phys. Rev. Lett.},
+  journal   = {Physical Review Letters},
   year      = {2003},
   volume    = {91},
   pages     = {220801},
@@ -299,10 +299,10 @@ Program title: Meep Catalogue identifier: AEFU\_v1\_0 Program summary URL:: http
 
 @Article{oelschlaeger2021,
   author     = {Oelschläger, M. and Reiche, D. and Egerland, C. H. and Busch, K. and Intravaia, F.},
-  journal    = {arXiv:2110.13635 [quant-ph]},
+  journal    = {Physical Review A},
   title      = {Electromagnetic viscosity in complex structured environments: from blackbody to quantum friction
 },
-  year       = {20212},
+  year       = {2022},
   month      = nov,
[…]
