### 2023-07-13 #5160 @chegerland
@lucydot Your PR has been merged, thank you very much! 😄
I also added the school to the PhD thesis entry.
