### 2023-07-14 #5160 @kyleniemeyer
Hi @chegerland, I'm doing some final checks on the paper before accepting.

To address the warning messages above, please update all the instances where you are using `\rm{xx}` for text in math mode; I believe the correct command is `\text{xx}` or `\textrm{xx}`.

For some other minor edits in the paper: our preferred citation style does not treat references as proper objects in sentences. You have a few instances with this (e.g., "refer to Ref. (Oelschlager 2019)" and similar on the first page). Can you revise those sentences to treat the references as footnotes instead? My suggested edits are:
- Replace `For a comprehensive review of quantum friction in various contexts as well as related effects in dynamical nonequilibrium, we refer, for example, to the reviews in Refs. [@volokitin2007;@dedkov2017;@reiche2022] and the references therein.` with `For a comprehensive review of quantum friction in various contexts as well as related effects in dynamical nonequilibrium, we refer, for example, to the reviews by @volokitin2007, @dedkov2017, and @reiche2022.`
- Replace `For details of the procedure, we refer to Ref. [@oelschlager2019].` with `@oelschlager2019 provides details of the procedure.`
